+++ src/actions/QueryActions.js
@@ -20,19 +20,6 @@ function loadError(error) {
   }
 }
 
-function generateCategories(data) {
-  return mapValues(data, val => map(val, 'key'))
-}
-
-function updateCategories(data) {
-  const categories = data.data ? generateCategories(data.data.aggregations) : []
-
-  return {
-    type: actionTypes.UPDATE_CATEGORIES,
-    categories
-  }
-}
-
 export function updateKeyword(keyword) {
   return {
     type: actionTypes.UPDATE_KEYWORD,
@@ -44,41 +31,6 @@ export function updatePageNum(pageNum) {
   return {
     type: actionTypes.UPDATE_PAGE_NUM,
     pageNum
-  }
-}
-
-export function updateSelectedFilter(selectedFilter, index) {
-  return {
-    type: actionTypes.UPDATE_SELECTED_FILTER,
-    index,
-    selectedFilter
-  }
-}
-
-export function updateSelectedValue(selectedValue, index) {
-  return {
-    type: actionTypes.UPDATE_SELECTED_FILTER_VALUE,
-    index,
-    selectedValue
-  }
-}
-
-export function addFilter() {
-  return {
-    type: actionTypes.ADD_FILTER
-  }
-}
-
-export function removeSelectedFilter(index) {
-  return {
-    type: actionTypes.REMOVE_SELECTED_FILTER,
-    index
-  }
-}
-
-export function removeAllFilters() {
-  return {
-    type: actionTypes.REMOVE_ALL_FILTERS
   }
 }
 
@@ -104,15 +56,5 @@ export function requestApi() {
         dispatch(receiveData())
         dispatch(loadError(error))
       })
-  }
-}
-
-export function getCategories() {
-  return (dispatch, getState) => {
-    const data = buildParams(getState().query)
-
-    return getData(data).then(
-      response => dispatch(updateCategories(response)),
-    )
   }
 }
